cm2.0 changes with ksmyplan-embedded building
git-svn-id: https://svn.kuali.org/repos/student/contrib/myplan/branches/ks-myplan-cm2.0@57430 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-myplan/myplan-api/src/main/java/org/kuali/student/myplan/academicplan/dto/PlanItemInfo.java
@@ -6,6 +6,7 @@ import org.kuali.student.r2.common.dto.TypeStateEntityInfo;
 import org.w3c.dom.Element;
 
 import javax.xml.bind.annotation.*;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -39,7 +40,7 @@ public class PlanItemInfo extends TypeStateEntityInfo implements PlanItem {
     private List<String> planPeriods;
 
     @XmlElement
-    private Float credit;
+    private BigDecimal credit;
 
     @XmlAnyElement
     private List<Element> _futureElements;
@@ -57,14 +58,14 @@ public class PlanItemInfo extends TypeStateEntityInfo implements PlanItem {
     public PlanItemInfo(PlanItem item) {
         super(item);
 
-        if(null != item) {
+        if (null != item) {
             this.id = item.getId();
             this.refObjectId = item.getRefObjectId();
             this.refObjectType = item.getRefObjectType();
             this.learningPlanId = item.getLearningPlanId();
 
-            if(null != item.getPlanPeriods()) {
-                for(String atpId : item.getPlanPeriods()) {
+            if (null != item.getPlanPeriods()) {
+                for (String atpId : item.getPlanPeriods()) {
                     this.planPeriods.add(atpId);
                 }
             }
@@ -122,11 +123,11 @@ public class PlanItemInfo extends TypeStateEntityInfo implements PlanItem {
         this.planPeriods = planPeriods;
     }
 
-     public Float getCredit() {
+    public BigDecimal getCredit() {
         return credit;
     }
 
-    public void setCredit(Float credit) {
-         this.credit = credit;
+    public void setCredit(BigDecimal credit) {
+        this.credit = credit;
     }
 }
